+++ packages/core/ios/RNSentrySDK.m
@@ -1,10 +1,6 @@
 #import "RNSentrySDK.h"
-#if __has_include(<RNSentry/RNSentry-Swift.h>)
-#    import <RNSentry/RNSentry-Swift.h>
-#else
-#    import "RNSentry-Swift.h"
-#endif
 #import "RNSentryStart.h"
+#import "RNSentrySwiftBridge.h"
 #import <Sentry/Sentry.h>
 
 static NSString *SENTRY_OPTIONS_RESOURCE_NAME = @"sentry.options";